"StreamAPI commonly asked questions updated"
